refactor: use cargo workspace
