+++ Cargo.toml
@@ -1,10 +1,3 @@
-[package]
-name = "tetris-in-rust"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-macroquad = "0.4"
-rand = "0.8.5"
+[workspace]
+members = ["tetris_game", "tetris_macroquad"]
+resolver = "2"
