--- src/lib.rs
@@ -1,7 +1,0 @@
-pub mod common;
-pub mod tetromino;
-pub mod tetris;
-pub mod ui;
-pub mod play;
-
-pub use play::play;
