+++ tetris_game/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "tetris_game"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+rand = "0.8.5"
