+++ tetris_game/src/lib.rs
@@ -1,0 +1,6 @@
+mod common;
+mod tetromino;
+mod tetris;
+
+pub use common::{Position, Color, Button, GamePad, GameUI};
+pub use tetris::Tetris;
