+++ tetris_macroquad/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "tetris_macroquad"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+macroquad = "0.4"
+tetris_game = { path = "../tetris_game" }
