+++ tetris_macroquad/src/lib.rs
@@ -1,0 +1,4 @@
+mod ui;
+mod play;
+
+pub use play::play_game;
